lugarden_universal/lu_the_poet/dify_output/requirements.txt
# Dify API 测试脚本依赖包

# HTTP 请求库
requests>=2.28.0

# 环境变量管理
python-dotenv>=1.0.0

# JSON 处理（Python内置，无需额外安装）
# datetime（Python内置，无需额外安装）
